rename module, misc. adjustments
