--- hr_timesheet_project_adj/README.rst
@@ -1,52 +1,0 @@
-========================
-HR Timesheet Project Adj
-========================
-
-.. !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
-   !! This file is generated by oca-gen-addon-readme !!
-   !! changes will be overwritten.                   !!
-   !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
-
-.. |badge1| image:: https://img.shields.io/badge/maturity-Beta-yellow.png
-    :target: https://odoo-community.org/page/development-status
-    :alt: Beta
-.. |badge2| image:: https://img.shields.io/badge/licence-AGPL--3-blue.png
-    :target: http://www.gnu.org/licenses/agpl-3.0-standalone.html
-    :alt: License: AGPL-3
-.. |badge3| image:: https://img.shields.io/badge/github-qrtl%2Fserver--auth-lightgray.png?logo=github
-    :target: https://github.com/qrtl/server-auth/tree/16.0/hr_timesheet_project_adj
-    :alt: qrtl/server-auth
-
-|badge1| |badge2| |badge3| 
-
-This module adjusts to not change project when task of project is changed for created timesheet records.
-
-**Table of contents**
-
-.. contents::
-   :local:
-
-Bug Tracker
-===========
-
-Bugs are tracked on `GitHub Issues <https://github.com/qrtl/server-auth/issues>`_.
-In case of trouble, please check there if your issue has already been reported.
-If you spotted it first, help us smashing it by providing a detailed and welcomed
-`feedback <https://github.com/qrtl/server-auth/issues/new?body=module:%20hr_timesheet_project_adj%0Aversion:%2016.0%0A%0A**Steps%20to%20reproduce**%0A-%20...%0A%0A**Current%20behavior**%0A%0A**Expected%20behavior**>`_.
-
-Do not contact contributors directly about support or help with technical issues.
-
-Credits
-=======
-
-Authors
-~~~~~~~
-
-* Quartile Limited
-
-Maintainers
-~~~~~~~~~~~
-
-This module is part of the `qrtl/server-auth <https://github.com/qrtl/server-auth/tree/16.0/hr_timesheet_project_adj>`_ project on GitHub.
-
-You are welcome to contribute.
