--- hr_timesheet_project_adj/readme/DESCRIPTION.rst
@@ -1,1 +1,0 @@
-This module adjusts to not change project when task of project is changed for created timesheet records.
