+++ hr_timesheet_unpropagate_project/README.rst
@@ -1,0 +1,57 @@
+================================
+HR Timesheet Unpropagate Project
+================================
+
+.. !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
+   !! This file is generated by oca-gen-addon-readme !!
+   !! changes will be overwritten.                   !!
+   !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
+
+.. |badge1| image:: https://img.shields.io/badge/maturity-Beta-yellow.png
+    :target: https://odoo-community.org/page/development-status
+    :alt: Beta
+.. |badge2| image:: https://img.shields.io/badge/licence-AGPL--3-blue.png
+    :target: http://www.gnu.org/licenses/agpl-3.0-standalone.html
+    :alt: License: AGPL-3
+.. |badge3| image:: https://img.shields.io/badge/github-qrtl%2Fqrtl--custom-lightgray.png?logo=github
+    :target: https://github.com/qrtl/qrtl-custom/tree/16.0/hr_timesheet_unpropagate_project
+    :alt: qrtl/qrtl-custom
+
+|badge1| |badge2| |badge3| 
+
+This module changes the behavior of standard Odoo to ensure that when you reassign a
+task to a different project, any timesheets linked to that task are not automatically
+updated.
+
+This can be useful in situations where your operation assumes that timesheets may be
+linked to different projects than those of their tasks.
+
+**Table of contents**
+
+.. contents::
+   :local:
+
+Bug Tracker
+===========
+
+Bugs are tracked on `GitHub Issues <https://github.com/qrtl/qrtl-custom/issues>`_.
+In case of trouble, please check there if your issue has already been reported.
+If you spotted it first, help us smashing it by providing a detailed and welcomed
+`feedback <https://github.com/qrtl/qrtl-custom/issues/new?body=module:%20hr_timesheet_unpropagate_project%0Aversion:%2016.0%0A%0A**Steps%20to%20reproduce**%0A-%20...%0A%0A**Current%20behavior**%0A%0A**Expected%20behavior**>`_.
+
+Do not contact contributors directly about support or help with technical issues.
+
+Credits
+=======
+
+Authors
+~~~~~~~
+
+* Quartile Limited
+
+Maintainers
+~~~~~~~~~~~
+
+This module is part of the `qrtl/qrtl-custom <https://github.com/qrtl/qrtl-custom/tree/16.0/hr_timesheet_unpropagate_project>`_ project on GitHub.
+
+You are welcome to contribute.
