+++ hr_timesheet_unpropagate_project/__manifest__.py
@@ -1,7 +1,7 @@
 # Copyright 2023 Quartile Limited
 # License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl).
 {
-    "name": "HR Timesheet Project Adj",
+    "name": "HR Timesheet Unpropagate Project",
     "category": "Timesheet",
     "version": "16.0.1.0.0",
     "author": "Quartile Limited",
