+++ hr_timesheet_unpropagate_project/readme/DESCRIPTION.rst
@@ -1,0 +1,6 @@
+This module changes the behavior of standard Odoo to ensure that when you reassign a
+task to a different project, any timesheets linked to that task are not automatically
+updated.
+
+This can be useful in situations where your operation assumes that timesheets may be
+linked to different projects than those of their tasks.
