+++ hr_timesheet_unpropagate_project/static/description/index.html
@@ -4,7 +4,7 @@
 <head>
 <meta http-equiv="Content-Type" content="text/html; charset=utf-8" />
 <meta name="generator" content="Docutils: http://docutils.sourceforge.net/" />
-<title>HR Timesheet Project Adj</title>
+<title>HR Timesheet Unpropagate Project</title>
 <style type="text/css">
 
 /*
@@ -360,15 +360,19 @@ ul.auto-toc {
 </style>
 </head>
 <body>
-<div class="document" id="hr-timesheet-project-adj">
-<h1 class="title">HR Timesheet Project Adj</h1>
+<div class="document" id="hr-timesheet-unpropagate-project">
+<h1 class="title">HR Timesheet Unpropagate Project</h1>
 
 <!-- !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
 !! This file is generated by oca-gen-addon-readme !!
 !! changes will be overwritten.                   !!
 !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!! -->
-<p><a class="reference external" href="https://odoo-community.org/page/development-status"><img alt="Beta" src="https://img.shields.io/badge/maturity-Beta-yellow.png" /></a> <a class="reference external" href="http://www.gnu.org/licenses/agpl-3.0-standalone.html"><img alt="License: AGPL-3" src="https://img.shields.io/badge/licence-AGPL--3-blue.png" /></a> <a class="reference external" href="https://github.com/qrtl/server-auth/tree/16.0/hr_timesheet_project_adj"><img alt="qrtl/server-auth" src="https://img.shields.io/badge/github-qrtl%2Fserver--auth-lightgray.png?logo=github" /></a></p>
-<p>This module adjusts to not change project when task of project is changed for created timesheet records.</p>
+<p><a class="reference external" href="https://odoo-community.org/page/development-status"><img alt="Beta" src="https://img.shields.io/badge/maturity-Beta-yellow.png" /></a> <a class="reference external" href="http://www.gnu.org/licenses/agpl-3.0-standalone.html"><img alt="License: AGPL-3" src="https://img.shields.io/badge/licence-AGPL--3-blue.png" /></a> <a class="reference external" href="https://github.com/qrtl/qrtl-custom/tree/16.0/hr_timesheet_unpropagate_project"><img alt="qrtl/qrtl-custom" src="https://img.shields.io/badge/github-qrtl%2Fqrtl--custom-lightgray.png?logo=github" /></a></p>
+<p>This module changes the behavior of standard Odoo to ensure that when you reassign a
+task to a different project, any timesheets linked to that task are not automatically
+updated.</p>
+<p>This can be useful in situations where your operation assumes that timesheets may be
+linked to different projects than those of their tasks.</p>
 <p><strong>Table of contents</strong></p>
 <div class="contents local topic" id="contents">
 <ul class="simple">
@@ -382,10 +386,10 @@ ul.auto-toc {
 </div>
 <div class="section" id="bug-tracker">
 <h1><a class="toc-backref" href="#id1">Bug Tracker</a></h1>
-<p>Bugs are tracked on <a class="reference external" href="https://github.com/qrtl/server-auth/issues">GitHub Issues</a>.
+<p>Bugs are tracked on <a class="reference external" href="https://github.com/qrtl/qrtl-custom/issues">GitHub Issues</a>.
 In case of trouble, please check there if your issue has already been reported.
 If you spotted it first, help us smashing it by providing a detailed and welcomed
-<a class="reference external" href="https://github.com/qrtl/server-auth/issues/new?body=module:%20hr_timesheet_project_adj%0Aversion:%2016.0%0A%0A**Steps%20to%20reproduce**%0A-%20...%0A%0A**Current%20behavior**%0A%0A**Expected%20behavior**">feedback</a>.</p>
+<a class="reference external" href="https://github.com/qrtl/qrtl-custom/issues/new?body=module:%20hr_timesheet_unpropagate_project%0Aversion:%2016.0%0A%0A**Steps%20to%20reproduce**%0A-%20...%0A%0A**Current%20behavior**%0A%0A**Expected%20behavior**">feedback</a>.</p>
 <p>Do not contact contributors directly about support or help with technical issues.</p>
 </div>
 <div class="section" id="credits">
@@ -398,7 +402,7 @@ If you spotted it first, help us smashing it by providing a detailed and welcome
 </div>
 <div class="section" id="maintainers">
 <h2><a class="toc-backref" href="#id4">Maintainers</a></h2>
-<p>This module is part of the <a class="reference external" href="https://github.com/qrtl/server-auth/tree/16.0/hr_timesheet_project_adj">qrtl/server-auth</a> project on GitHub.</p>
+<p>This module is part of the <a class="reference external" href="https://github.com/qrtl/qrtl-custom/tree/16.0/hr_timesheet_unpropagate_project">qrtl/qrtl-custom</a> project on GitHub.</p>
 <p>You are welcome to contribute.</p>
 </div>
 </div>
