--- setup/hr_timesheet_project_adj/odoo/addons/hr_timesheet_project_adj
@@ -1,1 +1,0 @@
-../../../../hr_timesheet_project_adj
