+++ setup/hr_timesheet_unpropagate_project/odoo/addons/hr_timesheet_unpropagate_project
@@ -1,0 +1,1 @@
+../../../../hr_timesheet_unpropagate_project
